executors: move CmdExecutor based files to cmdexec dir
Now that CmdExecutor is fully established, move all of the CmdExecutor
specific parts out of sshexec into its own cmdexec. This needs a bunch
of package renames as well to keep it compiling.
